requirements.txt
fastapi 
uvicorn 
python-multipart 
jinja2
pandas
numpy
openai
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
python-dotenv
tqdm
